Move functions in gce_extensions.go.
* Move functions in gce_extensions.go into ir/gateway_converter.go so
  these helper functions are eaiser to locate.
